Separating out the parallel aggregator complexity into its own file
- intend to create a less complicated single-thread version, too
