+++ net.dougqh.jist/.settings/org.eclipse.pde.core.prefs
@@ -1,0 +1,5 @@
+#Thu Mar 07 08:51:39 EST 2013
+eclipse.preferences.version=1
+pluginProject.equinox=false
+pluginProject.extensions=false
+resolve.requirebundle=false
